Feat: add the ability to provide code snippets and delete projects

+++ src/components/Projects/index.jsx
@@ -2,19 +2,25 @@ import { useEffect, useState } from "react";
 
 import { IoTrashOutline } from "react-icons/io5";
 import { FaRegCopy, FaCheck } from "react-icons/fa";
-
 import axios from "axios";
 import { useUserStore } from "../../store/store";
 
 import Modal from "../Modal";
+import DeleteConfirmationModal from "../DeleteConfirmationModal";
+
+import generateUsePageViewTrackerCode from "../../apis";
+import validateProjectName from "../../utils/validationUtils";
 
 function Projects() {
   const { user } = useUserStore();
   const [projects, setProjects] = useState([]);
   const [projectName, setProjectName] = useState("");
+  const [projectNameError, setProjectNameError] = useState("");
   const [isModalOpen, setIsModalOpen] = useState(false);
-  const [apiKey, setApiKey] = useState("");
+  const [responseData, setResponseData] = useState({});
   const [isCopied, setIsCopied] = useState(false);
+  const [isDeleteModalOpen, setIsDeleteModalOpen] = useState(false);
+  const [projectToDelete, setProjectToDelete] = useState(null);
 
   useEffect(() => {
     async function getProjects() {
@@ -40,6 +46,13 @@ function Projects() {
   async function handleSubmitProject(e) {
     e.preventDefault();
 
+    const errorMessage = validateProjectName(projectName);
+
+    if (errorMessage) {
+      setProjectNameError(errorMessage);
+      return;
+    }
+
     try {
       const response = await axios.post(
         `${import.meta.env.VITE_SERVER_URL}/users/${user._id}/projects`,
@@ -54,25 +67,37 @@ function Projects() {
 
       setProjects([...projects, response.data]);
       setProjectName("");
-      setApiKey(response.data.apiKey);
+      setResponseData(response.data);
+      setIsCopied(false);
     } catch (error) {
       console.error(error);
     }
   }
 
-  async function handleDeleteProject(projectId) {
+  async function handleDeleteProject() {
+    if (!projectToDelete) return;
+
     try {
       await axios.delete(
-        `${import.meta.env.VITE_SERVER_URL}/users/${user._id}/projects/${projectId}`,
+        `${import.meta.env.VITE_SERVER_URL}/users/${user._id}/projects/${projectToDelete}`,
         {
           withCredentials: true,
         },
       );
 
-      setProjects(projects.filter((project) => project._id !== projectId));
+      setProjects(
+        projects.filter((project) => project._id !== projectToDelete),
+      );
+      setIsDeleteModalOpen(false);
+      setProjectToDelete(null);
     } catch (error) {
       console.log(error);
     }
+  }
+
+  function handleDeleteModal(projectId) {
+    setIsDeleteModalOpen(true);
+    setProjectToDelete(projectId);
   }
 
   function formatDate(dateString) {
@@ -90,18 +115,30 @@ function Projects() {
   }
 
   function handleCloseModal() {
-    setApiKey("");
+    setResponseData({});
     setIsModalOpen(false);
+    setProjectNameError("");
   }
 
-  async function handleCopyApiKey() {
-    try {
-      await navigator.clipboard.writeText(apiKey);
+  function handleProjectNameChange(e) {
+    const newName = e.target.value;
 
+    setProjectName(newName);
+
+    if (projectNameError) {
+      setProjectNameError("");
+    }
+  }
+
+  async function handleCopyCode() {
+    const codeSnippet = generateUsePageViewTrackerCode(responseData);
+
+    try {
+      await navigator.clipboard.writeText(codeSnippet);
       setIsCopied(true);
       setTimeout(() => setIsCopied(false), 3000);
     } catch (error) {
-      console.error("Failed to copy API key:", error);
+      console.error("Failed to copy the code:", error);
     }
   }
 
@@ -126,7 +163,7 @@ function Projects() {
                 <td className="text-center">{formatDate(project.createdAt)}</td>
                 <td className="text-center">
                   <button
-                    onClick={() => handleDeleteProject(project._id)}
+                    onClick={() => handleDeleteModal(project._id)}
                     aria-label="delete"
                   >
                     <IoTrashOutline />
@@ -144,7 +181,7 @@ function Projects() {
         + Create new project
       </button>
       <Modal isOpen={isModalOpen} onClose={handleCloseModal}>
-        {!apiKey && (
+        {!responseData.apiKey && (
           <form onSubmit={handleSubmitProject} className="space-y-4">
             <label
               htmlFor="projectName"
@@ -157,53 +194,71 @@ function Projects() {
               id="projectName"
               name="projectName"
               value={projectName}
-              onChange={(e) => setProjectName(e.target.value)}
+              onChange={handleProjectNameChange}
               className="mt-1 p-2 w-full border border-gray-300 rounded-md shadow-sm"
               placeholder="프로젝트 이름을 입력하세요"
               required
             />
+            {projectNameError && (
+              <p className="text-red-500 text-xs italic">{projectNameError}</p>
+            )}
             <button
               type="submit"
               className="px-4 py-2 bg-blue-500 text-white rounded hover:bg-blue-700"
             >
-              Api key 생성
+              Api key & Service Code 생성
             </button>
           </form>
         )}
-        {apiKey && (
+        {responseData.apiKey && (
           <>
             <h1 className="text-center text-lg font-semibold mb-4">
-              API Key generated
+              API Key generated & Service Code
             </h1>
             <div className="mb-15">
               이 비밀 키를 안전하고 접근하기 쉬운 곳에 보관하세요. 보안상의
               이유로 FlowCathcer 계정을 통해서는 다시 볼 수 없습니다. 이 비밀
               키를 분실하신 경우 새 키를 생성하셔야 합니다.
             </div>
+            <div className="mb-15">
+              userId, apiKey, projectId 는 모두 안전하게 보호해 주셔야 합니다.
+            </div>
+            <h2 className="text-center text-lg font-semibold mb-4">
+              프로젝트명: {responseData.projectName}
+            </h2>
             <div className="flex items-center space-x-3">
-              <p className="font-mono text-sm bg-gray-300 p-1 ml-10 mr-10 rounded">
-                {apiKey}
-              </p>
-              <button
-                onClick={handleCopyApiKey}
-                className="p-2 bg-green-500 hover:bg-green-600 active:bg-green-700 text-white rounded-md shadow focus:outline-none focus:ring flex items-center justify-center"
-              >
-                {isCopied ? (
-                  <>
-                    <FaCheck />
-                    <span className="ml-2 text-white text-sm">Copied</span>
-                  </>
-                ) : (
-                  <>
-                    <FaRegCopy />
-                    <span className="ml-2 text-sm">Copy code</span>
-                  </>
-                )}
-              </button>
+              <div className="w-full">
+                <textarea
+                  readOnly
+                  value={generateUsePageViewTrackerCode(responseData)}
+                  className="font-mono text-sm bg-gray-100 p-2 w-full h-150 border border-gray-300 rounded-md"
+                />
+                <button
+                  onClick={handleCopyCode}
+                  className="p-2 bg-green-500 hover:bg-green-600 active:bg-green-700 text-white rounded-md shadow focus:outline-none focus:ring flex items-center justify-center"
+                >
+                  {isCopied ? (
+                    <>
+                      <FaCheck />
+                      <span className="ml-2 text-white text-sm">Copied</span>
+                    </>
+                  ) : (
+                    <>
+                      <FaRegCopy />
+                      <span className="ml-2 text-sm">Copy code</span>
+                    </>
+                  )}
+                </button>
+              </div>
             </div>
           </>
         )}
       </Modal>
+      <DeleteConfirmationModal
+        isOpen={isDeleteModalOpen}
+        onClose={() => setIsDeleteModalOpen(false)}
+        onConfirm={handleDeleteProject}
+      />
     </div>
   );
 }
